BFCOption {BiocFileCache}	R Documentation

BFCOption These functions help get and set an R variable CACHE that controls the default caching location.

Description

BFCOption These functions help get and set an R variable CACHE that controls the default caching location.

Usage

setBFCOption(arg, value)

Arguments

arg
character(1) option to get or set

value
The value to be assigned to the designated option

Details

Currently the only supported option is CACHE. This controls the default location of the BiocFileCache caching directory. By default the value is established by tools::R_user_dir("BiocFileCache",which="cache"). This value can also be defaultly set by using system and global environment variables visible before the package is loaded. The variable that should be set if utilized is “BFC_CACHE”

Value

Value of request option or invisible successfully set option

Author(s)

Lori Shepherd

Examples

origPath = getBFCOption('CACHE')
setBFCOption('CACHE', "~/.myBFC") 

[Package BiocFileCache version 2.2.1 Index]