makeBiocFileCacheFromDataFrame {BiocFileCache}	R Documentation

Make BiocFileCache objects from an existing data.frame

Description

If there are a lot of resources being added this could take some time but if a cache is saved in a permanent location this should only have to be run once. The original data.frame must have the required columns 'rtype', 'fpath', and 'rpath'; See the vignette for more information on the expected information contained in these columns. Similarly, the optional columns 'rname', 'etag', 'last_modified_time', and 'expires' may be included. Any additional columns not listed as required or optional will be kept as an additional metadata table in the BiocFileCache database.

Usage

makeBiocFileCacheFromDataFrame(df, cache, actionLocal = c("move", "copy",
  "asis"), actionWeb = c("move", "copy"), metadataName, ..., ask = TRUE)

## S4 method for signature 'ANY'
makeBiocFileCacheFromDataFrame(df, cache,
  actionLocal = c("move", "copy", "asis"), actionWeb = c("move", "copy"),
  metadataName, ..., ask = TRUE)

Arguments

df
data.frame or tibble to convert

cache
character(1) On-disk location (directory path) of cache. For default location see user_cache_dir.

actionLocal
If local copy of file should be moved, copied or left in original location. See 'action' param of bfcadd.

actionWeb
If a local copy of a remote resource already exists, should the file be copied or moved to the cache. Locally downloaded remote resources must exist in the cache location.

metadataName
If there are additional columns of data in the original data.frame besides required BiocFileCache columns, this data will be added as a metadata table with this name.

...
additional arguments passed to 'file.copy()'.

ask
logical(1) Confirm creation of BiocFileCache.

Value

A BiocFileCache object

[Package BiocFileCache version 1.6.0 Index]